requirements.txt
# Core dependencies
aiohttp>=3.8.0
asyncio-mqtt>=0.10.0
click>=8.0.0
requests>=2.28.0
PyJWT>=2.6.0
pyyaml>=6.0
pathlib
ipaddress

# Enhanced CLI dependencies
rich>=13.0.0
questionary>=1.10.0
colorama>=0.4.6

# Kubernetes and Docker
kubernetes>=25.3.0
docker>=6.0.0

# Web interface
flask>=2.3.0
flask-socketio>=5.8.0

# Integrations
python-telegram-bot>=20.0
discord.py>=2.3.0

# Development and testing
pytest>=7.2.0
pytest-asyncio>=0.21.0
black>=22.12.0
flake8>=6.0.0

# Security and validation
cryptography>=40.0.0
certifi>=2022.12.7

# Performance monitoring
psutil>=5.9.0
memory-profiler>=0.61.0

# Optional enhancements
pandas>=1.5.0  # For advanced data analysis
matplotlib>=3.6.0  # For visualization
seaborn>=0.12.0  # For enhanced plots